+++ src/components/Footer/Footer.scss
@@ -22,8 +22,8 @@
   }
 
   &__newsletter-form {
-    margin-top: 42px;
     color: var(--color-white);
+    margin-top: 42px;
   }
 
   &__social-media-links {
