+++ src/components/Newsletter/Newsletter.scss
@@ -1,23 +1,23 @@
 .newsletter-modal {
+  align-items: center;
+  background: rgba($color: #000000, $alpha: 0.3);
+  display: flex;
+  height: 100%;
+  justify-content: center;
   position: fixed;
   width: 100%;
-  height: 100%;
-  background: rgba($color: #000000, $alpha: 0.3);
   z-index: 2;
-  display: flex;
-  justify-content: center;
-  align-items: center;
 
   &__content {
-    display: flex;
-    flex-direction: column;
-    text-align: center;
-    padding: 76px 50px;
-    max-width: 540px;
     background: var(--color-white);
     border-radius: 3px;
-    position: relative;
+    display: flex;
+    flex-direction: column;
     margin: 10px;
+    max-width: 540px;
+    padding: 76px 50px;
+    position: relative;
+    text-align: center;
 
     @media (max-width: 425px) {
       padding: 60px 15px;
@@ -29,8 +29,8 @@
 
     .Icon {
       position: absolute;
-      top: 17px;
       right: 17px;
+      top: 17px;
     }
   }
