+++ src/components/NewsletterForm/NewsletterForm.scss
@@ -8,8 +8,8 @@
   }
 
   input {
-    width: 240px;
     padding: 8px;
+    width: 240px;
 
     @media (max-width: 425px) {
       width: 100%;
